hotkeys: fix incorrect implementation of hotkeys.js
The problem with the previous implementation was that it called
combo handler and single key handler twice which isn't correct;
If Ctrl+E was pressed E should not be triggred, also it used keypress
event which did not capture meta key or and captured Ctrl and A in one
event which is much more convinent though since it doesn't capture Meta
we used keydown event here, and the previous capuring previously captured key
works just like a drop-in replacement.

The tests are also re-written since previously they were much more unreadable
and not easy to follow, and most importantly they were testing wrong behavoir which
gave a chace to rewrite them intead of modifing exsisting one.