reminder-app: add initial design for displaying reminders.
This is just basically not yet done version, which does not look ugly.
Currently, something looks off because we have not breaker between title
and content, once we add icons, and stuff this should be gone!

This is mostly not responsive, though making this responsive would be that
hard right now we just need to get the basic good implementation ready -- css
change are always required no matter what.