migration-tests: make them run in order to avoid error when running multiple test files.
If there were more than one migration test files, the test fails, beacuse
they would run in parallel, hence, when the one is running we reset the db
so the one thats running fails. This fixes it.

If you add a new test file, this will not work since, both will run in
parallel, when we want them to run in order.