app: add editing-ui js module.
This just shows and hide editing ui, current not being used anywhere.
There is also some more thing to fix before we can fix this, like
editing title is fine, but editing the reminder is currently not working
as expected. It scrolls downwards, so we might as well show a textarea with
border.